pyproject.toml
[project]
name = "signup-login"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.13",
    "anthropic>=0.54.0",
    "fastapi-mcp>=0.3.4",
    "fastapi-mcp-client>=0.4.0",
    "fastapi[standard]>=0.115.13",
    "google-genai>=1.22.0",
    "ipykernel>=6.29.5",
    "openai>=1.91.0",
    "passlib[bcrypt]>=1.7.4",
    "pika>=1.3.2",
    "pyjwt>=2.10.1",
    "pymongo[srv]>=4.13.2",
    "streamlit>=1.46.0",
    "websockets>=15.0.1",
]
